% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangle}
\alias{wrangle}
\title{Wrangle "Time Tracker" data}
\usage{
wrangle(data)
}
\arguments{
\item{data}{A data frame as produced with "Time Tracker".}
}
\value{
A data frame.
}
\description{
Wrangle "Time Tracker" data
}
\examples{
# wrangle(googlesheets4::read_sheet("your/url"))
}
